refactor(logging): member join/leave functions
Remove `logMemberJoined` and `logMemberLeft` from `LoggingModule`.
